backend/requirements.txt
imgcompare
pyppeteer
python-dotenv
pymongo
dnspython